several bug fixes for the optimistic mode
In reverse handler of a completion event, we were marking the
associated task as undone. This is wrong, task should not be undone
here, because reversing a completion event does not mean reversing the
execution of the event. A task is marked as done when it's executed
even though it's completion event has not arrived yet.

Also again in reverse handler of an completion event, we move messages
from shadow to real buffer. We need to undone forward dependencies of
the tasks that are associated with these messages but we are not doing
this.

handle_exec_event: no need to store bool pe state in the buffer
PE_removeFromCopyBuffer: deletion will be always in the end, no need
to search the whole buffer

handle_recv_rev_event: if wasBusy - no forward dep. undone is needed

Change-Id: I01eccc984b4c95483817b5170873dad295d6a4db